+++ ui/plugins/markdown/shared/core/Database.js
@@ -1,42 +1,74 @@
-var config = require('config');
-var { findWhere, uniq, pluck } = require('lodash');
-var { makeHref } = require('actions/RouteActions');
-var strHumanize = require('tinydoc/lib/utils/strHumanize');
+const config = require('config');
+const { assign, where, findWhere, uniq, pluck } = require('lodash');
+const { makeHref } = require('actions/RouteActions');
+const strHumanize = require('tinydoc/lib/utils/strHumanize');
+const arrayWrap = require('tinydoc/lib/utils/arrayWrap');
 
-var folders = {};
+const articles = config.database;
+let folders = [];
 
-module.exports = {
-  getCollectionTitle(collectionName) {
-    return findWhere(config.collections, { name: collectionName }).title;
+let Database = {
+  getTitle() {
+    return config.title;
   },
 
-  getArticleTitles(collection) {
-    return config.database[collection];
+  getArticles() {
+    return articles;
   },
 
-  get(collection, articleId) {
-    return findWhere(config.database[collection], { id: decodeURIComponent(articleId) });
+  get(articleId) {
+    return findWhere(articles, { id: decodeURIComponent(articleId) });
   },
 
-  getFolders(collection) {
-    if (!folders[collection]) {
-      folders[collection] = uniq( pluck(config.database[collection], 'folder') );
-    }
-
-    return folders[collection];
+  getFolders() {
+    return folders;
   },
 
   getLinkableEntities() {
-    return Object.keys(config.database).reduce(function(links, collection) {
-
-      config.database[collection].forEach(function(entry) {
-        links[entry.filePath] = {
-          href: makeHref(`${collection}.article`, { articleId: entry.id }),
-          title: `${strHumanize(collection)}: ${entry.title}`
-        };
-      });
+    return articles.reduce(function(links, entry) {
+      links[entry.filePath] = {
+        href: makeHref(`${config.name}.article`, { articleId: entry.id }),
+        title: `${strHumanize(config.title)}: ${entry.title}`
+      };
 
       return links;
     }, {});
   }
 };
+
+function createFolderConfig(folderPath) {
+  const folderConfig = findWhere(config.folders, { path: folderPath });
+
+  let folder = assign({}, folderConfig, { path: folderPath });
+
+  // generate a title
+  if (!folder.title) {
+    if (config.fullFolderTitles) {
+      folder.title = folderPath
+        .replace(config.commonPrefix, '')
+        .split('/')
+        .map(strHumanize)
+        .join(config.fullFolderTitleDelimiter)
+      ;
+    }
+    else {
+      const fragments = folderPath.split('/');
+      folder.title = strHumanize(fragments[fragments.length-1]);
+    }
+  }
+
+  return folder;
+}
+
+// generate folders
+config.database.forEach((article) => {
+  if (!findWhere(folders, { path: article.folder })) {
+    folders.push(createFolderConfig(article.folder));
+  }
+});
+
+folders.forEach(function(folder) {
+  folder.articles = where(articles, { folder: folder.path });
+});
+
+module.exports = Database;